Refactoring the structure to clean things up
