--- brainsets/utils/open_neuro_utils/__init__.py
@@ -1,23 +1,0 @@
-"""OpenNeuro utilities subpackage."""
-
-from .data_extraction import (
-    extract_brainset_description,
-    extract_channels,
-    extract_device_description,
-    extract_meas_date,
-    extract_session_description,
-    extract_signal,
-    extract_subject_description,
-    generate_train_valid_splits_one_epoch,
-)
-
-__all__ = [
-    "extract_brainset_description",
-    "extract_channels",
-    "extract_device_description",
-    "extract_meas_date",
-    "extract_session_description",
-    "extract_signal",
-    "extract_subject_description",
-    "generate_train_valid_splits_one_epoch",
-]
